feat(irs-edc-client): [TRI-1594] Extension of Policy Checker Routine - move policy model to edc client
