+++ irs-edc-client/src/main/java/org/eclipse/tractusx/irs/edc/client/policy/Constraints.java
@@ -21,7 +21,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
-package org.eclipse.tractusx.irs.policystore.models;
+package org.eclipse.tractusx.irs.edc.client.policy;
 
 import java.util.List;
 
@@ -30,15 +30,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 /**
- * A Permission object gather PolicyType with list of Constraints and its relation
+ * A Constraint containing two collections of Constraints
  */
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class Permission {
+@SuppressWarnings("PMD.ShortVariable")
+public class Constraints {
 
-    private PolicyType action;
-    private List<Constraints> constraints;
-
+    private List<Constraint> and;
+    private List<Constraint> or;
 
 }
