+++ irs-edc-client/src/main/java/org/eclipse/tractusx/irs/edc/client/policy/OperatorType.java
@@ -21,7 +21,7 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
-package org.eclipse.tractusx.irs.policystore.models;
+package org.eclipse.tractusx.irs.edc.client.policy;
 
 import java.util.NoSuchElementException;
 import java.util.stream.Stream;
