+++ irs-edc-client/src/main/java/org/eclipse/tractusx/irs/edc/client/policy/Permission.java
@@ -21,9 +21,8 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
-package org.eclipse.tractusx.irs.policystore.models;
+package org.eclipse.tractusx.irs.edc.client.policy;
 
-import java.time.OffsetDateTime;
 import java.util.List;
 
 import lombok.AllArgsConstructor;
@@ -31,20 +30,15 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 /**
- * A stored policy object.
+ * A Permission object gather PolicyType with list of Constraints and its relation
  */
 @Getter
 @AllArgsConstructor
 @NoArgsConstructor
-public class Policy {
+public class Permission {
 
-    private String policyId;
-    private OffsetDateTime createdOn;
-    private OffsetDateTime validUntil;
-    private List<Permission> permissions;
+    private PolicyType action;
+    private List<Constraints> constraints;
 
-    public Policy update(final OffsetDateTime validUntil) {
-        this.validUntil = validUntil;
-        return this;
-    }
+
 }
