+++ irs-edc-client/src/main/java/org/eclipse/tractusx/irs/edc/client/policy/Policy.java
@@ -21,13 +21,30 @@
  *
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
-package org.eclipse.tractusx.irs.policystore.models;
+package org.eclipse.tractusx.irs.edc.client.policy;
+
+import java.time.OffsetDateTime;
+import java.util.List;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
 
 /**
- * A PolicyType object use in Permission
+ * A stored policy object.
  */
-public enum PolicyType {
+@Getter
+@AllArgsConstructor
+@NoArgsConstructor
+public class Policy {
 
-    ACCESS, USE
+    private String policyId;
+    private OffsetDateTime createdOn;
+    private OffsetDateTime validUntil;
+    private List<Permission> permissions;
 
+    public Policy update(final OffsetDateTime validUntil) {
+        this.validUntil = validUntil;
+        return this;
+    }
 }
