+++ irs-edc-client/src/main/java/org/eclipse/tractusx/irs/edc/client/policy/PolicyCheckerService.java
@@ -82,7 +82,7 @@ public class PolicyCheckerService {
         return policyStore.getAcceptedPolicies()
                           .stream()
                           .filter(p -> p.validUntil().isAfter(OffsetDateTime.now()))
-                          .map(AcceptedPolicy::policyId)
+                          .map(acceptedPolicy -> acceptedPolicy.policy().getPolicyId())
                           .flatMap(this::addEncodedVersion)
                           .toList();
     }
