+++ irs-edc-client/src/test/java/org/eclipse/tractusx/irs/edc/client/policy/AcceptedPoliciesProviderTest.java
@@ -62,6 +62,6 @@ class AcceptedPoliciesProviderTest {
     }
     @NotNull
     private static AcceptedPolicy policy() {
-        return new AcceptedPolicy("", OffsetDateTime.now());
+        return new AcceptedPolicy(new Policy(), OffsetDateTime.now());
     }
 }
