+++ irs-edc-client/src/test/java/org/eclipse/tractusx/irs/edc/client/policy/PolicyCheckerServiceTest.java
@@ -32,6 +32,7 @@ import static org.eclipse.tractusx.irs.edc.client.testutil.TestMother.createXOne
 import static org.mockito.Mockito.when;
 
 import java.time.OffsetDateTime;
+import java.util.Collections;
 import java.util.List;
 
 import org.eclipse.edc.policy.model.Policy;
@@ -52,8 +53,8 @@ class PolicyCheckerServiceTest {
 
     @BeforeEach
     void setUp() {
-        final var policyList = List.of(new AcceptedPolicy("ID 3.0 Trace", OffsetDateTime.now().plusYears(1)),
-                new AcceptedPolicy("FrameworkAgreement.traceability", OffsetDateTime.now().plusYears(1)));
+        final var policyList = List.of(new AcceptedPolicy(policy("ID 3.0 Trace"), OffsetDateTime.now().plusYears(1)),
+                new AcceptedPolicy(policy("FrameworkAgreement.traceability"), OffsetDateTime.now().plusYears(1)));
         when(policyStore.getAcceptedPolicies()).thenReturn(policyList);
         final List<String> leftOperands = List.of("PURPOSE");
         final List<String> rightOperands = List.of("active");
@@ -101,8 +102,8 @@ class PolicyCheckerServiceTest {
     @Test
     void shouldConfirmValidPolicyWhenWildcardIsSet() {
         // given
-        final var policyList = List.of(new AcceptedPolicy("ID 3.0 Trace", OffsetDateTime.now().plusYears(1)),
-                new AcceptedPolicy("*", OffsetDateTime.now().plusYears(1)));
+        final var policyList = List.of(new AcceptedPolicy(policy("ID 3.0 Trace"), OffsetDateTime.now().plusYears(1)),
+                new AcceptedPolicy(policy("*"), OffsetDateTime.now().plusYears(1)));
         when(policyStore.getAcceptedPolicies()).thenReturn(policyList);
         Policy policy = createAtomicConstraintPolicy("FrameworkAgreement.traceability", "active");
         // when
@@ -115,7 +116,7 @@ class PolicyCheckerServiceTest {
     @Test
     void shouldRejectWhenWildcardIsPartOfPolicy() {
         // given
-        final var policyList = List.of(new AcceptedPolicy("Policy*", OffsetDateTime.now().plusYears(1)));
+        final var policyList = List.of(new AcceptedPolicy(policy("Policy*"), OffsetDateTime.now().plusYears(1)));
         when(policyStore.getAcceptedPolicies()).thenReturn(policyList);
         Policy policy = createAtomicConstraintPolicy("FrameworkAgreement.traceability", "active");
         // when
@@ -129,9 +130,9 @@ class PolicyCheckerServiceTest {
     void shouldValidateAndConstraints() {
         // given
         final var policyList = List.of(
-                new AcceptedPolicy("FrameworkAgreement.traceability", OffsetDateTime.now().plusYears(1)),
-                new AcceptedPolicy("Membership", OffsetDateTime.now().plusYears(1)),
-                new AcceptedPolicy("ID 3.1 Trace", OffsetDateTime.now().plusYears(1)));
+                new AcceptedPolicy(policy("FrameworkAgreement.traceability"), OffsetDateTime.now().plusYears(1)),
+                new AcceptedPolicy(policy("Membership"), OffsetDateTime.now().plusYears(1)),
+                new AcceptedPolicy(policy("ID 3.1 Trace"), OffsetDateTime.now().plusYears(1)));
         when(policyStore.getAcceptedPolicies()).thenReturn(policyList);
         Policy policy = createAndConstraintPolicy(
                 List.of(createAtomicConstraint("FrameworkAgreement.traceability", "active"),
@@ -148,8 +149,8 @@ class PolicyCheckerServiceTest {
     void shouldRejectAndConstraintsWhenOnlyOneMatch() {
         // given
         final var policyList = List.of(
-                new AcceptedPolicy("FrameworkAgreement.traceability", OffsetDateTime.now().plusYears(1)),
-                new AcceptedPolicy("FrameworkAgreement.dismantler", OffsetDateTime.now().plusYears(1)));
+                new AcceptedPolicy(policy("FrameworkAgreement.traceability"), OffsetDateTime.now().plusYears(1)),
+                new AcceptedPolicy(policy("FrameworkAgreement.dismantler"), OffsetDateTime.now().plusYears(1)));
         when(policyStore.getAcceptedPolicies()).thenReturn(policyList);
         Policy policy = createAndConstraintPolicy(
                 List.of(createAtomicConstraint("FrameworkAgreement.traceability", "active"),
@@ -165,8 +166,8 @@ class PolicyCheckerServiceTest {
     void shouldValidateOrConstraints() {
         // given
         final var policyList = List.of(
-                new AcceptedPolicy("FrameworkAgreement.traceability", OffsetDateTime.now().plusYears(1)),
-                new AcceptedPolicy("FrameworkAgreement.dismantler", OffsetDateTime.now().plusYears(1)));
+                new AcceptedPolicy(policy("FrameworkAgreement.traceability"), OffsetDateTime.now().plusYears(1)),
+                new AcceptedPolicy(policy("FrameworkAgreement.dismantler"), OffsetDateTime.now().plusYears(1)));
         when(policyStore.getAcceptedPolicies()).thenReturn(policyList);
         Policy policy = createOrConstraintPolicy(
                 List.of(createAtomicConstraint("FrameworkAgreement.traceability", "active"),
@@ -181,8 +182,8 @@ class PolicyCheckerServiceTest {
     @Test
     void shouldRejectOrConstraintsWhenNoneMatch() {
         // given
-        final var policyList = List.of(new AcceptedPolicy("FrameworkAgreement.test", OffsetDateTime.now().plusYears(1)),
-                new AcceptedPolicy("FrameworkAgreement.dismantler", OffsetDateTime.now().plusYears(1)));
+        final var policyList = List.of(new AcceptedPolicy(policy("FrameworkAgreement.test"), OffsetDateTime.now().plusYears(1)),
+                new AcceptedPolicy(policy("FrameworkAgreement.dismantler"), OffsetDateTime.now().plusYears(1)));
         when(policyStore.getAcceptedPolicies()).thenReturn(policyList);
         Policy policy = createAndConstraintPolicy(
                 List.of(createAtomicConstraint("FrameworkAgreement.traceability", "active"),
@@ -198,8 +199,8 @@ class PolicyCheckerServiceTest {
     void shouldValidateXOneConstraints() {
         // given
         final var policyList = List.of(
-                new AcceptedPolicy("FrameworkAgreement.traceability", OffsetDateTime.now().plusYears(1)),
-                new AcceptedPolicy("FrameworkAgreement.dismantler", OffsetDateTime.now().plusYears(1)));
+                new AcceptedPolicy(policy("FrameworkAgreement.traceability"), OffsetDateTime.now().plusYears(1)),
+                new AcceptedPolicy(policy("FrameworkAgreement.dismantler"), OffsetDateTime.now().plusYears(1)));
         when(policyStore.getAcceptedPolicies()).thenReturn(policyList);
 
         Policy policy = createXOneConstraintPolicy(
@@ -215,8 +216,8 @@ class PolicyCheckerServiceTest {
     @Test
     void shouldRejectXOneConstraintsWhenNoneMatch() {
         // given
-        final var policyList = List.of(new AcceptedPolicy("FrameworkAgreement.test", OffsetDateTime.now().plusYears(1)),
-                new AcceptedPolicy("FrameworkAgreement.dismantler", OffsetDateTime.now().plusYears(1)));
+        final var policyList = List.of(new AcceptedPolicy(policy("FrameworkAgreement.test"), OffsetDateTime.now().plusYears(1)),
+                new AcceptedPolicy(policy("FrameworkAgreement.dismantler"), OffsetDateTime.now().plusYears(1)));
         when(policyStore.getAcceptedPolicies()).thenReturn(policyList);
         Policy policy = createXOneConstraintPolicy(
                 List.of(createAtomicConstraint("FrameworkAgreement.traceability", "active"),
@@ -232,8 +233,8 @@ class PolicyCheckerServiceTest {
     void shouldRejectXOneConstraintsWhenMoreThanOneMatch() {
         // given
         final var policyList = List.of(
-                new AcceptedPolicy("FrameworkAgreement.traceability", OffsetDateTime.now().plusYears(1)),
-                new AcceptedPolicy("Membership", OffsetDateTime.now().plusYears(1)));
+                new AcceptedPolicy(policy("FrameworkAgreement.traceability"), OffsetDateTime.now().plusYears(1)),
+                new AcceptedPolicy(policy("Membership"), OffsetDateTime.now().plusYears(1)));
         when(policyStore.getAcceptedPolicies()).thenReturn(policyList);
         Policy policy = createXOneConstraintPolicy(
                 List.of(createAtomicConstraint("FrameworkAgreement.traceability", "active"),
@@ -249,9 +250,9 @@ class PolicyCheckerServiceTest {
     void shouldAcceptMultipleRightOperands() {
         // given
         final var policyList = List.of(
-                new AcceptedPolicy("FrameworkAgreement.traceability", OffsetDateTime.now().plusYears(1)),
-                new AcceptedPolicy("FrameworkAgreement.dismantler", OffsetDateTime.now().plusYears(1)),
-                new AcceptedPolicy("Membership", OffsetDateTime.now().plusYears(1)));
+                new AcceptedPolicy(policy("FrameworkAgreement.traceability"), OffsetDateTime.now().plusYears(1)),
+                new AcceptedPolicy(policy("FrameworkAgreement.dismantler"), OffsetDateTime.now().plusYears(1)),
+                new AcceptedPolicy(policy("Membership"), OffsetDateTime.now().plusYears(1)));
         when(policyStore.getAcceptedPolicies()).thenReturn(policyList);
         final PolicyCheckerService testee = new PolicyCheckerService(policyStore, List.of("active", "inactive"),
                 List.of());
@@ -265,5 +266,14 @@ class PolicyCheckerServiceTest {
 
         // then
         assertThat(result).isTrue();
+    }
+
+    private org.eclipse.tractusx.irs.edc.client.policy.Policy policy(String policyId) {
+        return new org.eclipse.tractusx.irs.edc.client.policy.Policy(
+                policyId,
+                OffsetDateTime.now().plusYears(1),
+                OffsetDateTime.now().plusYears(1),
+                Collections.emptyList()
+        );
     }
 }
